MOB-25 Log 파일 생성 작업
