+++ build.gradle
@@ -15,6 +15,8 @@ configurations {
     compileOnly {
         extendsFrom annotationProcessor
     }
+
+
 }
 
 repositories {
@@ -30,8 +32,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
-    implementation 'mysql:mysql-connector-java:8.0.32'
 
+    implementation 'mysql:mysql-connector-java:8.0.32'
     compileOnly 'org.projectlombok:lombok'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
     runtimeOnly 'com.h2database:h2'
