+++ logs/http/http.2023-09-19.log
@@ -1,0 +1,39 @@
+2023-09-19 22:26:40.464 [http-nio-8080-exec-2] DEBUG o.a.coyote.http11.Http11InputBuffer - Before fill(): parsingHeader: [true], parsingRequestLine: [true], parsingRequestLinePhase: [0], parsingRequestLineStart: [0], byteBuffer.position(): [0], byteBuffer.limit(): [0], end: [0]
+2023-09-19 22:26:40.464 [http-nio-8080-exec-1] DEBUG o.a.coyote.http11.Http11InputBuffer - Before fill(): parsingHeader: [true], parsingRequestLine: [true], parsingRequestLinePhase: [0], parsingRequestLineStart: [0], byteBuffer.position(): [0], byteBuffer.limit(): [0], end: [0]
+2023-09-19 22:26:40.464 [http-nio-8080-exec-2] DEBUG o.a.coyote.http11.Http11InputBuffer - Received [POST /api/login HTTP/1.1
+Content-Type: application/json
+User-Agent: PostmanRuntime/7.33.0
+Accept: */*
+Postman-Token: 9b8455e8-fb23-498e-9b9b-d8f2049199e8
+Host: localhost:8080
+Accept-Encoding: gzip, deflate, br
+Connection: keep-alive
+Content-Length: 80
+
+
+{
+    "id":"test999",
+    "password":"test999",
+    "nickname":"test999"
+}]
+2023-09-19 22:26:40.696 [http-nio-8080-exec-2] DEBUG o.a.coyote.http11.Http11InputBuffer - Before fill(): parsingHeader: [true], parsingRequestLine: [true], parsingRequestLinePhase: [0], parsingRequestLineStart: [0], byteBuffer.position(): [0], byteBuffer.limit(): [0], end: [263]
+2023-09-19 22:26:40.696 [http-nio-8080-exec-2] DEBUG o.a.coyote.http11.Http11InputBuffer - Received []
+2023-09-19 22:26:59.977 [http-nio-8080-exec-3] DEBUG o.a.coyote.http11.Http11InputBuffer - Before fill(): parsingHeader: [true], parsingRequestLine: [true], parsingRequestLinePhase: [0], parsingRequestLineStart: [0], byteBuffer.position(): [0], byteBuffer.limit(): [0], end: [263]
+2023-09-19 22:26:59.977 [http-nio-8080-exec-3] DEBUG o.a.coyote.http11.Http11InputBuffer - Received [POST /api/login HTTP/1.1
+Content-Type: application/json
+User-Agent: PostmanRuntime/7.33.0
+Accept: */*
+Postman-Token: dbae89a8-1ee8-4b33-9ac1-ca2d9c1a31e5
+Host: localhost:8080
+Accept-Encoding: gzip, deflate, br
+Connection: keep-alive
+Content-Length: 80
+
+
+{
+    "id":"test999",
+    "password":"test999",
+    "nickname":"test999"
+}]
+2023-09-19 22:26:59.988 [http-nio-8080-exec-3] DEBUG o.a.coyote.http11.Http11InputBuffer - Before fill(): parsingHeader: [true], parsingRequestLine: [true], parsingRequestLinePhase: [0], parsingRequestLineStart: [0], byteBuffer.position(): [0], byteBuffer.limit(): [0], end: [263]
+2023-09-19 22:26:59.988 [http-nio-8080-exec-3] DEBUG o.a.coyote.http11.Http11InputBuffer - Received []
