+++ logs/sql/sql.2023-09-19.log
@@ -1,0 +1,20 @@
+2023-09-19 22:26:40.655 [http-nio-8080-exec-2] DEBUG org.hibernate.SQL - 
+    select
+        u1_0.id,
+        u1_0.nickname,
+        u1_0.password 
+    from
+        user u1_0 
+    where
+        u1_0.id=?
+2023-09-19 22:26:40.659 [http-nio-8080-exec-2] TRACE org.hibernate.orm.jdbc.bind - binding parameter [1] as [VARCHAR] - [test999]
+2023-09-19 22:26:59.981 [http-nio-8080-exec-3] DEBUG org.hibernate.SQL - 
+    select
+        u1_0.id,
+        u1_0.nickname,
+        u1_0.password 
+    from
+        user u1_0 
+    where
+        u1_0.id=?
+2023-09-19 22:26:59.982 [http-nio-8080-exec-3] TRACE org.hibernate.orm.jdbc.bind - binding parameter [1] as [VARCHAR] - [test999]
