+++ src/main/java/com/example/mswp/config/SecurityConfig.java
@@ -4,8 +4,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
+
+
 @Configuration
 public class SecurityConfig {
+
 
     @Bean
     public BCryptPasswordEncoder passwordEncoder() {
