+++ src/main/java/com/example/mswp/controller/UserController.java
@@ -4,13 +4,20 @@ import com.example.mswp.dto.UserDto;
 import com.example.mswp.entity.User;
 import com.example.mswp.mapping.UserMapping;
 import com.example.mswp.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpRequest;
 import org.springframework.web.bind.annotation.*;
 
 import javax.swing.text.html.Option;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @RestController
 @RequestMapping("/api")
@@ -19,6 +26,7 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    private Logger logger = LoggerFactory.getLogger(UserController.class);
     @GetMapping("/test")
     public String test() {
         return "Success!";
@@ -32,6 +40,18 @@ public class UserController {
     @PostMapping("/info")
     public Optional<UserMapping> info(@RequestBody UserDto userDto) {
         return userService.loadUser(userDto);
+    }
+
+    @PostMapping ("/register")
+    public Map SignUpMethod(HttpServletRequest request, @RequestBody UserDto userDto) {
+        System.out.println(request.getRemoteAddr());
+
+        Map<String, Object> responsedata = new HashMap<>();
+
+        String result = this.userService.SignUpMethod(userDto);
+        responsedata.put("sc",result);
+
+        return responsedata;
     }
 
 }
