+++ src/main/java/com/example/mswp/service/UserService.java
@@ -41,4 +41,22 @@ public class UserService {
     public Optional<UserMapping> loadUser(UserDto userDto) {
         return userRepository.findAllById(userDto.getId());
     }
+
+    public String SignUpMethod(UserDto userDto){
+        //패스워드 암호화
+        userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
+        // 아이디 찾는 메소드
+        Optional<User> userTest = userRepository.findById(userDto.getId());
+
+        String response = "400";
+
+        //유저 id가 db에 있는지 확인
+        if (userTest.isEmpty()){
+            User newUser = userDto.toEntity();
+            userRepository.save(newUser);
+            response = "200";
+        }
+        return response;
+    };
+
 }
