+++ src/main/resources/application.properties
@@ -12,3 +12,27 @@ spring.jpa.hibernate.ddl-auto=update
 
 #JPA ???? Hibernate? ????? ??? SQL? ??? ??
 spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.apache.coyote.http11=DEBUG
+
+
+#?? ?? ??
+
+logging.level.org.hibernate=info
+
+#logging.level.org.hibernate.SQL=debug
+#logging.level.org.hibernate.orm.jdbc.bind=trace
+
+#logging.level.org.hibernate.stat=debug
+#logging.level.org.hibernate.SQL_SLOW=debug
+
+#logging.level.org.hibernate.cache=debug
+#logging.level.org.springframework.web=debug
+
+#tomcat ?? http ?? log ??
+server.tomcat.basedir=.
+server.tomcat.accesslog.enabled=true
+server.tomcat.accesslog.pattern=%{yyyy-MM-dd HH:mm:ss.SSS}t\u0009%s\u0009%r\u0009%{User-Agent}i\u0009%{Referer}i\u0009%a\u0009%b
+server.tomcat.accesslog.directory=./logs/request
+server.tomcat.accesslog.prefix=req
+server.tomcat.accesslog.suffix=.log
+
