+++ src/main/resources/logback-spring.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+    <include resource="org/springframework/boot/logging/logback/base.xml"/>
+
+
+    <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
+        <file>./logs/main/myapp.log</file>
+        <encoder>
+            <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+            <!-- daily rollover -->
+            <fileNamePattern>./logs/main/myapp.%d{yyyy-MM-dd:ss.SSS}.log</fileNamePattern>
+            <maxHistory>300</maxHistory>
+            <!-- keep 30 days' worth of history -->
+            <!--<maxHistory>30</maxHistory>  -->
+        </rollingPolicy>
+    </appender>
+
+    <appender name="HTTP11" class="ch.qos.logback.core.rolling.RollingFileAppender">
+
+
+        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+            <!-- daily rollover -->
+            <fileNamePattern>./logs/http/http.%d{yyyy-MM-dd}.log</fileNamePattern>
+            <maxHistory>300</maxHistory>
+            <!-- keep 30 days' worth of history -->
+            <!--<maxHistory>30</maxHistory>  -->
+        </rollingPolicy>
+        <encoder>
+            <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <appender name="SQL" class="ch.qos.logback.core.rolling.RollingFileAppender">
+
+
+        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+            <!-- daily rollover -->
+            <fileNamePattern>./logs/sql/sql.%d{yyyy-MM-dd}.log</fileNamePattern>
+            <maxHistory>300</maxHistory>
+            <!-- keep 30 days' worth of history -->
+            <!--<maxHistory>30</maxHistory>  -->
+        </rollingPolicy>
+        <encoder>
+            <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+
+
+    <logger name = "org.apache.coyote.http11.Http11InputBuffer" level="DEBUG" additivity="false" ><appender-ref ref="HTTP11"/></logger>
+
+    <logger name = "org.hibernate.SQL" level="debug" additivity="false" ><appender-ref ref="SQL"/></logger>
+    <logger name = "org.hibernate.orm.jdbc.bind" level="trace" additivity="false" ><appender-ref ref="SQL"/></logger>
+
+    <root level="INFO">
+        <!--<appender-ref ref="FILE" />-->
+    </root>
+
+
+
+</configuration>
